% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{characterOrParameters-class}
\alias{characterOrParameters-class}
\alias{characterOrParameters}
\title{Class "characterOrParameters"}
\description{
A simple union class of \code{character} and \code{\linkS4class{parameters}}.
Objects will be created internally whenever necessary and the user should
not need to explicitly interact with this class.
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\examples{

showClass("characterOrParameters")

}
\keyword{classes}
