% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/IO.R
\name{checkOffset}
\alias{checkOffset}
\title{Fix the offset when its values recorded in header and TEXT don't agree}
\usage{
checkOffset(offsets, x, ignore.text.offset = FALSE, ...)
}
\arguments{
\item{offsets}{the named vector returned by \code{findOffsets}}

\item{x}{the text segmented returned by \code{readFCStext}}

\item{ignore.text.offset}{whether to ignore the offset info stored in TEXT segment}

\item{...}{not used.}
}
\value{
the updated offsets
}
\description{
Fix the offset when its values recorded in header and TEXT don't agree
}

