% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{collapse_desc}
\alias{collapse_desc}
\title{Coerce the list of the keywords into a character 
Also flatten spillover matrix into a string}
\usage{
collapse_desc(d, collapse.spill = TRUE)
}
\arguments{
\item{d}{a named list of keywords}

\item{collapse.spill}{whether to flatten spillover matrix to a string}
}
\value{
a list of strings
}
\description{
Coerce the list of the keywords into a character 
Also flatten spillover matrix into a string
}
\examples{
data(GvHD)
fr <- GvHD[[1]]
collapse_desc(keyword(fr))
}
