% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{complementFilter-class}
\alias{complementFilter-class}
\alias{complementFilter}
\alias{show,complementFilter-method}
\title{Class complementFilter}
\description{
This class represents the logical complement of a single filter, which is 
itself a filter that can be incorporated in to further set operations. 
\code{complementFilter}s are constructed using the prefix unary set operator 
\code{"!"} with a single filter operand.
}
\section{Slots}{

\describe{
\item{\code{filters}}{Object of class \code{"list"}, containing the component filters.}

\item{\code{filterId}}{Object of class \code{"character"}
referencing the filter applied.}
}}

\section{Extends}{


Class \code{"\linkS4class{filter}"}, directly.
}

\seealso{
\code{\link[flowCore:filter-methods]{filter}}, \code{\linkS4class{setOperationFilter}}

Other setOperationFilter classes: \code{\link{intersectFilter-class}},
  \code{\link{setOperationFilter-class}},
  \code{\link{subsetFilter-class}},
  \code{\link{unionFilter-class}}
}
\author{
B. Ellis
}
\concept{setOperationFilter classes}
\keyword{classes}
