% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoExploreR_wrappers.R
\name{CytoExploreR_exports}
\alias{CytoExploreR_exports}
\alias{CytoExploreR_.estimateLogicle}
\title{CytoExploreR exports}
\usage{
CytoExploreR_.estimateLogicle(x, channels, ...)
}
\description{
Exported wrappers of internal functions for use by CytoExploreR
}
\keyword{internal}
