% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median-logicle-transform.R
\name{estimateMedianLogicle}
\alias{estimateMedianLogicle}
\title{Estimates a common logicle transformation for a flowSet.}
\usage{
estimateMedianLogicle(flow_set, channels, m = 4.5, q = 0.05)
}
\arguments{
\item{flow_set}{object of class 'flowSet'}

\item{channels}{character vector of channels to transform}

\item{m}{TODO -- default value from .lgclTrans}

\item{q}{quantile}
}
\value{
TODO
}
\description{
Of the negative values for each channel specified, the median of the specified
quantiles are used.
}
