\name{filter-and-methods}
\docType{methods}
\alias{\%\&\%-methods}
\alias{\%\&\%}
\alias{\%\&\%,ANY-method}

\title{ Methods for Function \%\&\% in Package `flowCore'}
\description{
 This operator is used to construct
}
\section{Methods}{
  \describe{
    \item{}{ ~~describe this method here }
}}
\author{B. Ellis}
\keyword{methods}
