\name{filterDetails-methods}
\docType{methods}
\alias{filterDetails-methods}
\alias{filterDetails}
\alias{filterDetails,filterResult,missing-method}
\alias{filterDetails,filterResult,ANY-method}
\alias{filterDetails<-,}
\alias{filterDetails<-, filterResult,character,setOperationFilter-method}
\alias{filterDetails<-, filterResult, character,filter-method}
\alias{filterDetails<-, filterResult, character,ANY-method}

\title{ Methods for Function filterDetails in Package flowCore}
\description{
  Methods for function \code{filterDetails} in Package flowCore
}
\section{Methods}{
\describe{

\item{result = "filterResult", filterId = "missing"}{ ~~describe this method here }

\item{result = "filterResult", filterId = "ANY"}{ ~~describe this method here }
}}
\keyword{methods}

