\name{flowCore-package}
\alias{flowCore-package}
\alias{flowCore}
\docType{package}
\title{
  
Provides S4 data structures and basic infrastructure and functions to
deal with flow cytometry data. 

}
\description{

  Define important flow cytometry data classes: \code{\link{flowFrame}},
  \code{\link{flowSet}} and their acdcessors.
  
  Provide important transformation, filter, gating, workflow, and summary
  functions for flow cytometry data analysis.
  
  Most of flow cytometry related Bioconductor packages (such as
  flowStats, flowFP, flowQ, flowViz, flowMerge, flowClust) are heavily
  depended on this package.
}

\details{ \tabular{ll}{
    Package: \tab flowCore
    \cr Type: \tab Package
    \cr Version: \tab 1.11.19
    \cr Date: \tab 2009-09-15
    \cr License: \tab Artistic-2.0\cr }

}
\author{
  Maintainer: Florian Hahne <fhahne@fhcrc.org>
  Authors: B. Ellis, P. Haaland, F. Hahne, N. Le Meur, N. Gopalakrishnan
}
\references{

}

\keyword{ package }
\seealso{
\code{\link[flowCore:flowCore-package]{flowCore}}
}
\examples{
## examples go here
}
