% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSet-accessors.R
\name{flowSet_to_list}
\alias{flowSet_to_list}
\title{Convert a flowSet to a list of flowFrames}
\usage{
flowSet_to_list(fs)
}
\arguments{
\item{fs}{a flowSet}
}
\value{
a list of flowFrames
}
\description{
This is a simple helper function for splitting a \code{\link[flowCore:flowSet-class]{flowSet}}
in to a list of its constituent \code{\link[flowCore:flowFrame-class]{flowFrames}}.
}
