% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{intersectFilter-class}
\alias{intersectFilter-class}
\alias{intersectFilter}
\alias{show,intersectFilter-method}
\title{Class intersectFilter}
\description{
This class represents the intersection of two filters, which is itself a filter
that can be incorporated in to further set operations. \code{intersectFilter}s
are constructed using the binary set operator \code{"&"} with operands consisting
of a single filter or list of filters.
}
\section{Slots}{

\describe{
\item{\code{filters}}{Object of class \code{"list"}, containing the component filters.}

\item{\code{filterId}}{Object of class \code{"character"}
referencing the filter applied.}
}}

\section{Extends}{


Class \code{"\linkS4class{filter}"}, directly.
}

\seealso{
\code{\link[flowCore:filter-methods]{filter}}, \code{\linkS4class{setOperationFilter}}

Other setOperationFilter classes: 
\code{\link{complementFilter-class}},
\code{\link{setOperationFilter-class}},
\code{\link{subsetFilter-class}},
\code{\link{unionFilter-class}}
}
\author{
B. Ellis
}
\concept{setOperationFilter classes}
\keyword{classes}
