% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-logicleZ.R
\name{inverseLogicleZTransform}
\alias{inverseLogicleZTransform}
\title{Computes the inverse of the transform defined by the 'logicleZTransform'
function or the transformList generated by 'estimateLogicleZ' function}
\usage{
inverseLogicleZTransform(trans,transformationId, ...)
}
\arguments{
\item{trans}{An object of class 'transform' created using the
'logicleZTransform' function or class 'transformList' created by
'estimateLogicleZ'.  The parameters w, t, m, a for calculating the inverse
are obtained from the 'trans' input passed to the 'inverseLogicleZTransform'
function.}

\item{transformationId}{A name to assigned to the inverse transformation.
Used by the transform routines.}

\item{...}{not used.}
}
\description{
inverseLogicleTransform can be use to compute the inverse of the Logicle
transformation. The parameters w, t, m, a for calculating the inverse are
obtained from the 'trans' input passed to the 'inverseLogicleTransform'
function. (The inverseLogicleTransform method makes use of the C++
implementation of the inverse logicle transform contributed by Wayne Moore
et al.)
}
\examples{

# GvHD is not a recent dataset to show the interest of this transformation

}
\references{
Parks D.R., Roederer M., Moore W.A.(2006) A new "logicle"
display method avoids deceptive effects of logarithmic scaling for low
signals and compensated data. CytometryA, 96(6):541-51.
}
\seealso{
\code{\link[flowCore]{logicleZTransform}}
}
\keyword{methods}
