% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{multiRangeGate-class}
\alias{multiRangeGate-class}
\alias{multiRangeGate}
\alias{summary,multiRangeGate-method}
\alias{show,multiRangeGate-method}
\title{Multirange Gate class}
\usage{
multiRangeGate(ranges, filterId="defaultMultiRangeGate")
}
\arguments{
\item{filterId}{An optional parameter that sets the \code{filterId} of this
gate. The object can later be identified by this name.}

\item{ranges}{A definition of the gate. This can be a list of min,max ranges
(see the prototype).}
}
\value{
Returns a \code{\link{multiRangeGate}} object for use in filtering
\code{\link{flowFrame}}s or other flow cytometry objects.
}
\description{
Multirange Gate class
}
