% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{nullParameter-class}
\alias{nullParameter-class}
\alias{nullParameter}
\title{Class "nullParameter"}
\description{
A class used internally for coercing transforms to characters for a return
value when a coercion cannot be performed. The user should never need to
interact with this class.
}
\section{Objects from the Class}{
 Objects will be created internally whenever
necessary and this should not be of any concern to the user.
}

\keyword{classes}
