\name{print.filterSummary}
\alias{print.filterSummary}
\alias{summary}
\title{Summarize the results after applying a filter}
\description{
 Summarize the results after applying a filter on the \code{flowFrame} or
 \code{flowSet} object.
}
\usage{
summary(x, ...)
}

\arguments{
  \item{x}{object of \code{filterResult} class}
  \item{\dots}{optional arguments for the summary function}
}

\value{
 number and percentage of selected features (cells) in the population out of the number of
 features first recorded in the original dataset.
}

\author{B. Ellis}

\seealso{\code{filterResults}}
\examples{

}

\keyword{manip}
