% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{squareroot-class}
\alias{squareroot-class}
\alias{squareroot}
\alias{eval,squareroot,missing-method}
\title{Class "squareroot"}
\description{
Square root transform class, which represents a transformation defined by the 
function \deqn{f(parameter,a)= \sqrt{ |{\frac{parameter}{a}|}}}
}
\section{Slots}{

\describe{
\item{\code{.Data}}{Object of class \code{"function"}}

\item{\code{a}}{Object of class \code{"numeric"} -- non-zero multiplicative 
constant}

\item{\code{parameters}}{Object of class \code{"transformation"} -- flow 
parameter to be transformed.}

\item{\code{transformationId}}{Object of class \code{"character"} -- unique 
ID to reference the transformation.}
}}

\note{
The squareroot transformation object can be evaluated using the eval
method by passing the data frame as an argument.The transformed parameters
are returned as a column vector. (See example below)
}
\section{Objects from the Class}{
 Objects can be created by calls to the
constructor \code{squareroot(parameters,a,transformationId)}
}

\section{Extends}{

Class \code{"\linkS4class{singleParameterTransform}"}, directly.

Class \code{"\linkS4class{transform}"}, by class "singleParameterTransform", distance 2.

Class \code{"\linkS4class{transformation}"}, by class "singleParameterTransform", distance 3.

Class \code{"\linkS4class{characterOrTransformation}"}, by class "singleParameterTransform", distance 4.
}

\examples{

  dat <- read.FCS(system.file("extdata","0877408774.B08",
  package="flowCore"))
  sqrt1<-squareroot(parameters="FSC-H",a=2,transformationId="sqrt1")
  transOut<-eval(sqrt1)(exprs(dat))

}
\references{
Gating-ML Candidate Recommendation for Gating Description in
Flow Cytometry
}
\seealso{
dg1polynomial, ratio, quadratic

Other mathematical transform classes: \code{\link{EHtrans-class}},
  \code{\link{asinht-class}},
  \code{\link{asinhtGml2-class}},
  \code{\link{dg1polynomial-class}},
  \code{\link{exponential-class}},
  \code{\link{hyperlog-class}},
  \code{\link{hyperlogtGml2-class}},
  \code{\link{invsplitscale-class}},
  \code{\link{lintGml2-class}},
  \code{\link{logarithm-class}},
  \code{\link{logicletGml2-class}},
  \code{\link{logtGml2-class}},
  \code{\link{quadratic-class}}, \code{\link{ratio-class}},
  \code{\link{ratiotGml2-class}},
  \code{\link{sinht-class}},
  \code{\link{splitscale-class}},
  \code{\link{unitytransform-class}}
}
\author{
Gopalakrishnan N, F.Hahne
}
\concept{mathematical transform classes}
\keyword{classes}
