\name{transformation-class}
\docType{class}
\alias{transformation-class}

\title{Class "transformation"  }
\description{A virtual class to abtract transformations.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Extends}{
Class \code{"\linkS4class{characterOrTransformation}"}, directly.
}
\section{Methods}{
No methods defined with class "transformation" in the signature.
}
\author{N. Gopalakrishnan}



\keyword{classes}
