% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{transformReference-class}
\alias{transformReference-class}
\alias{transformReference}
\alias{parameters,transformReference-method}
\alias{eval,transformReference,missing-method}
\title{Class "transformReference"}
\description{
Class allowing for reference of transforms, for instance as parameters.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{The list of references.}

\item{\code{searchEnv}}{The environment into which the reference points.}

\item{\code{transformationId}}{The name of the transformation.}
}}

\section{Objects from the Class}{
 Objects will be created internally whenever
necessary and this should not be of any concern to the user.
}

\section{Extends}{

Class \code{"\linkS4class{transform}"}, directly.
Class \code{"\linkS4class{transformation}"}, by class "transform", distance 2.
Class \code{"\linkS4class{characterOrTransformation}"}, by class "transform", distance 3.
}

\author{
N. Gopalakrishnan
}
\keyword{classes}
