% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{unionFilter-class}
\alias{unionFilter-class}
\alias{unionFilter}
\alias{show,unionFilter-method}
\title{Class unionFilter}
\description{
This class represents the union of two filters, which is itself a filter
that can be incorporated in to further set operations. \code{unionFilter}s
are constructed using the binary set operator \code{"|"} with operands
consisting of a single \code{filter} or list of \code{filters}.
}
\section{Slots}{

\describe{
\item{\code{filters}}{Object of class \code{"list"}, containing the component filters.}

\item{\code{filterId}}{Object of class \code{"character"}
referencing the filter applied.}
}}

\section{Extends}{


Class \code{"\linkS4class{filter}"}, directly.
}

\seealso{
\code{\link[flowCore:filter-methods]{filter}}, \code{\linkS4class{setOperationFilter}}

Other setOperationFilter classes: \code{\link{complementFilter-class}},
  \code{\link{intersectFilter-class}},
  \code{\link{setOperationFilter-class}},
  \code{\link{subsetFilter-class}}
}
\author{
B. Ellis
}
\concept{setOperationFilter classes}
\keyword{classes}
