% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowFrame-accessors.R
\name{updateTransformKeywords}
\alias{updateTransformKeywords}
\title{modify description to reflect the transformation
Involve inserting/updating 'transformation' and flowCore_$PnRmax keywords}
\usage{
updateTransformKeywords(fr)
}
\arguments{
\item{fr}{flowFrame}
}
\value{
updated description slot
}
\description{
modify description to reflect the transformation
Involve inserting/updating 'transformation' and flowCore_$PnRmax keywords
}
