% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{validFilters}
\alias{validFilters}
\title{Check if all filters in a filters matches same paramters}
\usage{
validFilters(flist)
}
\arguments{
\item{flist}{a filters object}
}
\value{
TRUE or FALSE
}
\description{
Check if all filters in a filters matches same paramters
}
