% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{concreteFilter-class}
\alias{concreteFilter-class}
\alias{concreteFilter}
\title{Class "concreteFilter"}
\description{
The \code{concreteFilter} serves as a base class for all filters that
actually implement a filtering process. At the moment this includes all
filters except \code{\linkS4class{filterReference}}, the only non-concrete
filter at present.
}
\section{Slots}{

\describe{
\item{\code{filterId}}{The identifier associated with this class.}
}}

\section{Objects from the Class}{
 Objects of this class should never be
created directly. It serves only as a point of inheritance.
}

\section{Extends}{

Class \code{"\linkS4class{filter}"}, directly.
}

\seealso{
\code{\linkS4class{parameterFilter}}
}
\author{
B. Ellis
}
\keyword{classes}
