% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{filterReference-class}
\alias{filterReference-class}
\alias{filterReference}
\alias{filterReference,environment,character-method}
\alias{summary,filterReference-method}
\alias{length,filterReference-method}
\alias{show,filterReference-method}
\alias{eval,filterReference,missing-method}
\title{Class filterReference}
\description{
A reference to another filter inside a reference. Users should generally not
be aware that they are using this class, but it is used heavily by
\code{"\linkS4class{filterSet}"} classes.
}
\section{Slots}{

\describe{
\item{\code{name}}{The R name of the referenced filter.}

\item{\code{env}}{The environment where the filter must live.}

\item{\code{filterId}}{The filterId, not really used since you always resolve.}
}}

\section{Objects from the Class}{
 Objects are generally not created by users
so there is no constructor function.
}

\section{Extends}{


Class \code{"\linkS4class{filter}"}, directly.
}

\seealso{
\code{"\linkS4class{filterSet}"}
}
\author{
B. Ellis
}
\keyword{classes}
