% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowCore.R
\docType{package}
\name{flowCore-package}
\alias{flowCore-package}
\alias{flowCore}
\title{flowCore: Basic structures for flow cytometry data}
\description{
Provides S4 data structures and basic infrastructure and functions to deal
with flow cytometry data.
}
\details{
Define important flow cytometry data classes:
\code{\link[flowCore:flowFrame-class]{flowFrame}},
\code{\link[flowCore:flowSet-class]{flowSet}} and their accessors.

Provide important transformation, filter, gating, workflow, and summary
functions for flow cytometry data analysis.

Most of flow cytometry related Bioconductor packages (such as flowStats,
flowFP, flowQ, flowViz, flowMerge, flowClust) are heavily dependent on this
package.

\tabular{ll}{ Package: \tab flowCore \cr Type: \tab Package \cr Version:
\tab 1.11.20 \cr Date: \tab 2009-09-16 \cr License: \tab Artistic-2.0\cr }
}
\author{
Maintainer: Florian Hahne <fhahne@fhcrc.org>

Authors: B. Ellis, P. Haaland, F. Hahne, N. Le Meur, N. Gopalakrishnan
}
\keyword{package}
