\name{identifier}
\alias{identifier}
\alias{identifier,filterResult}

\title{Retrieve the GUID of a flowFrame}
\description{Retrieve the GUID or globally unique identifier of a
  \code{\link[flowCore]{flowframe} that was generated by the cytometer.}}

\usage{
identifier(object)
}

\arguments{
  \item{object}{Object of class \code{\link[flowCore]{flowFrame}}.}
}

\details{GUID or Globally Unique Identifier is a pseudo-random number
  used in software applications. While each generated GUID is not
  guaranteed to be unique, the total number of unique keys (2\^128)
  is so large that the probability of the same number being generated
  twice is very small.

  Note that if no GUID has been recorded along with the FCS file, the
  name of the file is returned.
}

\value{Character vector representing the GUID or the name of the file.
}


\author{N. Le Meur}

\examples{
 samp <- read.FCS(system.file("extdata","0877408774.B08", package="flowCore"))
 identifier(samp)

}
\keyword{methods}  

