% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-logicleZ.R
\name{inverseLogicleZTransform}
\alias{inverseLogicleZTransform}
\title{Computes the inverse of the transform defined by the 'logicleZTransform'
function or the transformList generated by 'estimateLogicleZ' function}
\usage{
inverseLogicleZTransform(trans, transformationId, ...)
}
\arguments{
\item{trans}{An object of class 'transform' created using the
'logicleZTransform' function or class 'transformList' created by
'estimateLogicleZ'.}

\item{transformationId}{A name to assigned to the inverse transformation.
Used by the transform routines.}

\item{...}{not used.}
}
\description{
inverseLogicleZTransform can be use to compute the inverse of the LogicleZ
transformation. The parameters w, t, m, a for calculating the inverse are
obtained from the 'trans' input passed to the 'inverseLogicleZTransform'
function.
}
\examples{

# GvHD is not a recent dataset to show the interest of this transformation

}
\seealso{
\code{\link[flowCore]{logicleZTransform}}
}
\keyword{methods}
