% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-logicleZ.R
\name{logicleZTransform}
\alias{logicleZTransform}
\alias{estimateLogicleZ}
\title{Computes a transform using the 'logicle_transform' function but keeping zero
at zero}
\usage{
logicleZTransform(transformationId="defaultLogicleTransform", w = 0.5,
  t = 262144, m = 4.5, a = 0)

estimateLogicle(x, channels,...)
}
\arguments{
\item{transformationId}{A name to assign to the transformation. Used by the
transform/filter routines.}

\item{w}{w of logicleTransform.}

\item{t}{t of logicleTransform.}

\item{m}{m of logicleTransform.}

\item{a}{a of logicleTransform.}

\item{z}{logical bring back zero at zero.}

\item{n}{negative intensity threshold.}

\item{x}{Input flow frame for which the logicle transformations are to be
estimated.}

\item{channels}{channels or markers for which the logicle transformation is
to be estimated.}

\item{...}{other arguments:

  q: a numeric type specifying quantile value, default is 0.05}
}
\description{
The default logicle transformation transforms the zero intensity into a value
that is not zero. The current transformation removes this offset after the
standard logicle transformation in order to keep zero at zero. See Logicle
transformation \code{\link{logicleTransform}} for all parameters.
}
\details{
The estimateLogicle function estimates the parameter w (width of the linear
zone) from the negative (left) part of a marker's negative peak. It
calculates the 5th percentile of intensities below the zero threshold. The
estimateLogicleZ function defines the center of the negative peak and
subtracts this value from all intensities before performing the classic
estimate.
}
\examples{

# GvHD is not a recent dataset to show the interest of this transformation

}
\references{
Parks D.R., Roederer M., Moore W.A.(2006) A new "logicle" display
  method avoids deceptive effects of logarithmic scaling for low signals and
  compensated data. CytometryA, 96(6):541-51.
}
\seealso{
\code{\link[flowCore]{inverseLogicleZTransform}},
  \code{\link[flowCore]{estimateLogicle} }
}
\keyword{methods}
