% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{parameters-class}
\alias{parameters-class}
\title{Class "parameters"}
\description{
A representation of flow parameters that allows for referencing.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list of the individual parameters.}
}}

\section{Objects from the Class}{
 Objects will be created internally whenever
necessary and this should not be of any concern to the user.
}

\section{Extends}{

Class \code{"\linkS4class{list}"}, from data part.
Class \code{"\linkS4class{vector}"}, by class "list", distance 2.
}

\author{
Nishant Gopalakrishnan
}
\keyword{classes}
