\name{print.filterSummary}
\alias{print.filterSummary}
\alias{summary}
\title{Summarize the results after applying a filter}
\description{
 Summarize the results after applying a filter on the \code{flowFrame} or
 \code{flowSet} object.
}
\usage{
summary(object, ...)
}

\arguments{
  \item{object}{an object of class \code{filterSummary} class}
  \item{\dots}{optional arguments for the summary function}
}

\value{
a print function for a \code{filterSummary} which provides counts and percentages for a 
filtering operation.
}

\author{B. Ellis}

\seealso{\code{filterResult}}
\examples{

}

\keyword{manip}
