\name{setOperationFilter-class}
\docType{class}
\alias{setOperationFilter-class}

\title{Class "setOperationFilter"}
\description{}

\section{Slots}{
	 \describe{
    \item{\code{filters}:}{Object of class \code{"list"}}
    \item{\code{filterId}:}{Object of class \code{"character"} referencing the filter applied}
  }
}
\section{Extends}{
Class \code{"\linkS4class{filter}"}, directly.
}
\author{B. Ellis}

\seealso{
  \code{\link[flowCore]{filter}}
}
\keyword{classes}
