% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSet-accessors.R
\docType{methods}
\name{spillover-flowSet}
\alias{spillover-flowSet}
\alias{spillover}
\alias{spillover,flowSet-method}
\title{Compute a spillover matrix from a flowSet}
\usage{
\S4method{spillover}{flowSet}(x, unstained = NULL, patt = NULL, fsc = "FSC-A", 
                              ssc = "SSC-A", method = "median", 
                              stain_match = c("intensity", "ordered", "regexpr"),
                              useNormFilt=FALSE, pregate = FALSE, plot = FALSE, \dots)
}
\arguments{
\item{x}{A flowSet of compensation beads or cells}

\item{unstained}{The name of index of the unstained negative control}

\item{patt}{An optional regular expression defining which parameters should
be considered.}

\item{fsc}{The name or index of the forward scatter parameter}

\item{ssc}{The name or index of the side scatter parameter}

\item{method}{The statistic to use for calculation. Traditionally, this has
been the median so it is the default. The mean is sometimes more stable.}

\item{stain_match}{Determines how the stain channels are matched with the
compensation controls. See details.}

\item{useNormFilt}{Logical Indicating whether to apply a
\code{\link{norm2Filter}} first before computing the spillover}

\item{pregate}{logical. Should we pregate each channel before computing the
spillover matrix? By default, no.}

\item{plot}{logical. Plots the kernel density for each channel when
pregating. Displays the gate used. If \code{pregate} is set to \code{FALSE},
this argument is ignored.}

\item{\dots}{Additional arguments passed to
\code{\link[flowStats]{rangeGate}}.}
}
\value{
A matrix for each of the parameters
}
\description{
Spillover information for a particular experiment is often obtained by
running several tubes of beads or cells stained with a single color that can
then be used to determine a spillover matrix for use with
\code{\link{compensate}}.\cr\cr
When matching stain channels in \code{x} with the compensation controls, we
provide a few options. If \code{ordered}, we assume the ordering of the
channels in the flowSet object is the same as the ordering of the
compensation-control samples. IF \code{regexpr}, we use a regular expression
to match the channel names with the filenames of the compensation controls.
By default, we must "guess" based on the largest statistic for the
compensation control (i.e., the row).
}
\details{
The algorithm used is fairly simple. First, using the scatter parameters, we
restrict ourselves to the most closely clustered population to reduce the
amount of debris. The selected statistic is then calculated on all
appropriate parameters and the unstained values swept out of the matrix.
Every sample is then normalized to [0,1] with respect to the maximum value
of the sample, giving the spillover in terms of a proportion of the primary
channel intensity.
}
\references{
C. B. Bagwell \& E. G. Adams (1993). Fluorescence spectral
overlap compensation for any number of flow cytometry parameters. in: Annals
of the New York Academy of Sciences, 677:167-184.
}
\seealso{
\code{\link{compensate}}
}
\author{
B. Ellis
}
\keyword{methods}
