% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{transformation-class}
\alias{transformation-class}
\alias{transformation}
\title{Class "transformation"}
\description{
A virtual class to abstract transformations.
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\section{Extends}{

Class \code{"\linkS4class{characterOrTransformation}"}, directly.
}

\author{
N. Gopalakrishnan
}
\keyword{classes}
