\name{transformList-class}
\docType{class}
\alias{transformList-class}
\alias{colnames,transformList-method}

\title{Class "transformList"}
\description{Class "transformList"}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("transformList", ...)}.
	 ~~ describe objects here ~~ 
}
\section{Slots}{
	 \describe{
    \item{\code{transforms}:}{Object of class \code{"list"}}
  }
}
\section{Methods}{
  \item{colnames}{add description here}.
}

\author{B. Ellis}

\seealso{
	\code{\link[flowCore]{transform}}
}
\examples{
showClass("transformList")
}
\keyword{classes}
